requirements.txt
# requirements.txt
asyncio
pyrofork
tgcrypto
yt-dlp
python-dotenv==1.0.0
aiosqlite
asyncio
aiofiles
aiohttp
flask
